+++ CHANGELOG.md
@@ -9,6 +9,7 @@ New release of the HyperDbg Debugger.
 
 ### Added
 - The **!xsetbv** event command is added to execution of XSETBV instruction thanks to HyperDbg group members ([link](https://docs.hyperdbg.org/commands/extension-commands/xsetbv))
+- Number of blocked context switches in the '.switch' command ([link](https://docs.hyperdbg.org/commands/meta-commands/.switch))
 
 ### Changed
 - Non-volatile XMM registers are no longer saved/restored on VM-exit handler ([link](https://learn.microsoft.com/en-us/cpp/build/x64-software-conventions?view=msvc-170))